Moves event filters to own file for use outside of liquid; import them for use within liquid.js.
